plugins with a go.mod
