+++ tools/rosetta/plugins/osmosis/Makefile
@@ -1,0 +1,4 @@
+#!/usr/bin/make -f
+
+plugin:
+	go build -buildmode=plugin -o osmosis.so osmosis.go
